add-bookings добавлены БД, jpa, querydsl. Рефакторинг кода под новое состояние.

+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -18,27 +18,27 @@ public class ItemController {
     private final ItemService itemService;
 
     @PostMapping
-    public ItemDto create(@RequestHeader(USER_ID) int userId, @Validated(Create.class) @RequestBody ItemDto itemDto) {
+    public ItemDto create(@RequestHeader(USER_ID) Long userId, @Validated(Create.class) @RequestBody ItemDto itemDto) {
         return itemService.create(userId, itemDto);
     }
 
     @GetMapping
-    public List<ItemDto> getItems(@RequestHeader(USER_ID) int userId) {
+    public List<ItemDto> getItems(@RequestHeader(USER_ID) Long userId) {
         return itemService.getItems(userId);
     }
 
     @GetMapping("/{id}")
-    public ItemDto getItemById(@RequestHeader(USER_ID) int userId, @PathVariable int id) {
+    public ItemDto getItemById(@RequestHeader(USER_ID) Long userId, @PathVariable Long id) {
         return itemService.getItemById(userId, id);
     }
 
     @PatchMapping("/{id}")
-    public ItemDto update(@RequestHeader(USER_ID) int userId, @PathVariable int id, @Validated(Update.class) @RequestBody ItemDto itemDto) {
+    public ItemDto update(@RequestHeader(USER_ID) Long userId, @PathVariable Long id, @Validated(Update.class) @RequestBody ItemDto itemDto) {
         return itemService.update(userId, id, itemDto);
     }
 
     @GetMapping("/search")
-    public List<ItemDto> getItemByText(@RequestHeader(USER_ID) int userId, @RequestParam (value = "text") String text) {
+    public List<ItemDto> getItemByText(@RequestHeader(USER_ID) Long userId, @RequestParam (value = "text") String text) {
         return itemService.getItemByText(userId, text);
     }
 }
