+++ src/main/java/ru/practicum/shareit/item/ItemMapper.java
@@ -31,7 +31,7 @@ public class ItemMapper {
         return item;
     }
 
-    public Item createItemFromDto(int userId, ItemDto itemDto) {
+    public Item createItemFromDto(Long userId, ItemDto itemDto) {
         return Item.builder()
                 .id(itemDto.getId())
                 .name(itemDto.getName())
